feat: add webpack build
