+++ cv/src/index.html
@@ -3,24 +3,23 @@
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Kirill Hrudzinski</title>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,400;0,600;0,700;1,400;1,600&display=swap"
           rel="stylesheet">
-    <link rel="stylesheet" href="styles/style.css">
-    <link rel="stylesheet" href="styles/github-dark-dimmed.min.css">
+    <link rel="stylesheet" href="style.css">
 </head>
 <body>
 <div class="page-wrapper">
     <header class="header">
         <div class="picture-wrapper">
-            <img class="cv-picture" src="assets/img/kirill-hrudzinski.png" alt="kirill-hrudzinski">
+
         </div>
         <div class="heading-wrapper">
             <h1 class="h1 heading heading-h1">Kirill Hrudzinski</h1>
-            <h3 class="h3 heading heading-h3">Front End Developer</h3>
+            <h2 class="h2 heading heading-h2">Front End Developer</h2>
         </div>
     </header>
     <aside class="sidebar">
@@ -246,7 +245,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Front End developer / Webmaster (Remote)</h4>
+                        <h3 class="description-list__h3">Front End developer / Webmaster (Remote)</h3>
                         <p class="description-list__paragraph">Mostly I'm creating web pages and editing existing, parts
                             pages,
                             payment cart, user's cabinet,
@@ -296,7 +295,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Lead Software Support Specialist</h4>
+                        <h3 class="description-list__h3">Lead Software Support Specialist</h3>
                         <p class="description-list__paragraph">Support products based on BPM Pega</p>
                         <div class="technologies">
                             <p class="technologies__paragraph">Technologies what I used:</p>
@@ -320,7 +319,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Senior Software Support Engineer</h4>
+                        <h3 class="description-list__h3">Senior Software Support Engineer</h3>
                         <p class="description-list__paragraph">Implementation and support of a credit line for small and
                             medium
                             businesses based on BPM Pega,
@@ -360,7 +359,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">IT Specialist</h4>
+                        <h3 class="description-list__h3">IT Specialist</h3>
                         <p class="description-list__paragraph">Implementation and support of the software of the Bank;
                             Participation in projects on software
                             development of the Bank, rendering consultations and assistance to the Bank's units;
@@ -389,7 +388,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Support Engineer</h4>
+                        <h3 class="description-list__h3">Support Engineer</h3>
                         <p class="description-list__paragraph">Administration and servicing of ATMs NCR, Wincor Nixdorf,
                             self-service terminals, Lanecom,
                             Hypercom, terminal equipment, controversial operations in the above equipment</p>
@@ -418,7 +417,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Support Engineer</h4>
+                        <h3 class="description-list__h3">Support Engineer</h3>
                         <p>Maintenance of SoftClub software, "Client-Bank"; PTSS, ATM, Lanecom, Hypercom, terminal
                             equipment; Minor repair of peripheral and copying devices; administration; backup;
                             Consultation
@@ -456,7 +455,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Pre School(Javascript)</h4>
+                        <h3 class="description-list__h3">Pre School(Javascript)</h3>
                         <p>During education I've done a couple of tasks:</p>
                         <ul class="acquired-skills">
                             <li class="acquired-skills__item">
@@ -507,7 +506,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Angular & React</h4>
+                        <h3 class="description-list__h3">Angular & React</h3>
                         <p>During education I lerned:</p>
                         <ul class="acquired-skills">
                             <li class="acquired-skills__item">Work with ES6 and ES7 language standards and some design
@@ -558,7 +557,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Advanced level, Front-end</h4>
+                        <h3 class="description-list__h3">Advanced level, Front-end</h3>
                         <p>During education I lerned:</p>
                         <ul class="acquired-skills">
                             <li class="acquired-skills__item">To work effectively with web pages;</li>
@@ -604,7 +603,7 @@
                 </dt>
                 <dd class="description-list__description">
                     <article class="description-list__article">
-                        <h4 class="description-list__h4">Electronics Engineer</h4>
+                        <h3 class="description-list__h3">Electronics Engineer</h3>
                         <p>The topic of my thesis project: "Gain control of a photodetector"</p>
                     </article>
                 </dd>
@@ -793,10 +792,5 @@
 
     </footer>
 </div>
-
-
-<script src="js/script.js"></script>
-<script src="js/highlight.min.js"></script>
-<script>hljs.initHighlightingOnLoad();</script>
 </body>
 </html>
