+++ cv/src/js/script.js
@@ -1,0 +1,47 @@
+const sideBarWrapper = document.querySelector('.sidebar-wrapper');
+const html = document.querySelector('html');
+const body = document.querySelector('body');
+
+const toggleSidebar = (e) => {
+    html.classList.toggle('active');
+    sideBarWrapper.classList.remove('visible');
+};
+
+const removeVisible = () => {
+    if (window.innerWidth <= 1140 && !html.classList.contains('active')) {
+        sideBarWrapper.classList.remove('visible');
+    }
+};
+
+const checkVisible = () => {
+    if (window.innerWidth > 1140|| html.classList.contains('active')) {
+        sideBarWrapper.classList.add('visible')
+    } else {
+        sideBarWrapper.classList.remove('visible')
+    }
+};
+const removeActive = (e) => {
+    if (e.target.classList.contains('aside-list__link') && window.innerWidth <= 1140 || e.target === body && window.innerWidth <= 1140) {
+        sideBarWrapper.classList.remove('visible');
+        html.classList.remove('active');
+    }
+};
+const createImage = (src) => new Promise((res, rej) => {
+    const img = new Image();
+    img.onload = () => res(img);
+    img.onerror = rej;
+    img.src = src;
+});
+
+export {
+    toggleSidebar,
+    removeActive,
+    removeVisible,
+    checkVisible,
+    createImage
+}
+
+
+
+
+
